+++ src/Integrations/Payment/Converge/ConvergeValidation.php
@@ -74,4 +74,14 @@ class ConvergeValidation extends Component
 //
 //        throw new \RuntimeException('Failed to load currency ISO codes.');
 //    }
+
+    public function validateLuhn($number)
+    {
+        $str = '';
+        foreach (array_reverse(str_split($number)) as $i => $c) {
+            $str .= $i % 2 ? $c * 2 : $c;
+        }
+
+        return array_sum(str_split($str)) % 10 === 0;
+    }
 }