+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Form Builder Integrations Changelog
 
+## 1.0.3 - 2019-01-02
+
+### Improved
+- Added friendly validation for CC number (if used as PlainText), stripped white spaces
+- Added friendly validation for Amount field
+
 ## 1.0.2 - 2018-12-20
 
 ### Fixed